ptn test refactor: move rdv permutation fn
